Move Qt dialog code out of data_acquisition into app/
data_acquisition must stay import-safe in headless environments.

- data_acquisition/gui_dialogs.py now contains only DialogRequest
  (pure dataclass, no Qt dependency)
- app/gui_dialogs.py (new) holds _build_device_error_message and
  show_device_error_dialog — all PyQt5 imports live here
- action_manager._prompt_user_quit_action: remove Qt fallback; auto-abort
  with a warning log when on_user_prompt is not wired
- Remove now-unused sys import and stale commented-out PyQt5 import
  from action_manager
